+++ AndroidTest15/app/src/main/java/com/example/ryota/androidtest15/TestDialogFragment.java
@@ -1,0 +1,35 @@
+package com.example.ryota.androidtest15;
+
+import android.annotation.SuppressLint;
+import android.app.Dialog;
+import android.app.DialogFragment;
+import android.content.DialogInterface;
+import android.os.Bundle;
+import android.support.v7.app.AlertDialog;
+import android.util.Log;
+
+@SuppressLint("ValidFragment")
+class TestDialogFragment extends DialogFragment {
+    @Override
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
+        // Use the Builder class for convenient dialog construction
+        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+        builder.setMessage("テストダイアログ")
+                .setPositiveButton("OK", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        // OK button pressed
+                        Log.i("AlertDialog1","ok");
+                    }
+                })
+                .setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        // OK button pressed
+                        Log.i("AlertDialog2","cancel");
+                    }
+                });
+        // Create the AlertDialog object and return it
+        return builder.create();
+    }
+}
